+++ src/tui_menus/game_ended.rs
@@ -257,11 +257,11 @@ impl<T: Write> Application<T> {
                             [
                                 (
                                     "Ctrl+Alt+L",
-                                    "Re-load from savefile (overwrites current data!)",
+                                    "Reload app from savefile (overwrites current data!)",
                                 ),
                                 (
                                     "Ctrl+Alt+S",
-                                    "Do savefile storage (respects save preferences)",
+                                    "Perform savefile store (respects save preferences)",
                                 ),
                                 ("Ctrl+C", "Exit program (respects save preferences)"),
                             ]