+++ src/tui_menus/new_game.rs
@@ -309,8 +309,8 @@ impl<T: Write> Application<T> {
                             ("Alt+↓/↑ Alt+j/k", "Adjust initial fall delay of Custom mode multiplicatively"),
                             ("Alt+Enter", "View game save as replay"),
                             ("Ctrl+U", "Unlock all game modes"),
-                            ("Ctrl+Alt+L", "Re-load from savefile (overwrites current data!)"),
-                            ("Ctrl+Alt+S", "Do savefile storage (respects save preferences)"),
+                            ("Ctrl+Alt+L", "Reload app from savefile (overwrites current data!)"),
+                            ("Ctrl+Alt+S", "Perform savefile store (respects save preferences)"),
                             ("Ctrl+C", "Exit program (respects save preferences)"),
                         ].into_iter().map(|(lhs,rhs)| (lhs.to_owned(), rhs.to_owned())).collect()),
                     ];